requirements.txt
alembic==1.7.1
attrs==21.2.0
blessed==1.17.6
botocore==1.21.36
certifi==2021.5.30
chalice==1.24.2
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
idna==3.2
importlib-resources==5.2.2
inquirer==2.7.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
Mako==1.1.5
MarkupSafe==2.0.1
mypy-extensions==0.4.3
psycopg2-binary==2.9.1
python-dateutil==2.8.2
python-editor==1.0.4
PyYAML==5.4.1
readchar==2.0.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.23
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
zipp==3.5.0
